% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LT}
\alias{LT}
\title{Mortality data}
\format{
A matrix with 38 columns and 17 rows.
}
\usage{
LT
}
\description{
A dataset containing mortality rates for Algerian men by 5-age groups over the 1977-2014 period.
}
\examples{
data(LT)
}
\keyword{datasets}
