% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2q.R
\name{getqx}
\alias{getqx}
\title{getqx}
\usage{
getqx(Qx, nag)
}
\arguments{
\item{Qx}{Five-ages mortality rates which can be a vector created using  or a column of a numerical matrix}

\item{nag}{number of age groups}
}
\value{
qx age-specific mortality rates

lx a vector containing the age evolution of survivorship

dx a vector containing the theoretical deaths occured at age x

qxtl age specific mortality rates interpolated using the Lagrange method

qxtk age specific mortality rates interpolated using the Karup-king method

jonct_age the age where qxk and qxl have been joined
}
\description{
It interpolate the age specific mortality rates
}
\examples{
getqx(c(0.12, seq(0.05, 0.8,by=0.05)), 17)

}
\author{
Farid FLICI
}
